Renames services for consistency
