+++ src/sentry/hybridcloud/rpc_services/region_organization_provisioning/service.py
@@ -1,14 +1,16 @@
 from abc import abstractmethod
 from typing import cast
 
-from sentry.hybridcloud.rpc_services.organization_provisioning import RpcOrganizationSlugReservation
+from sentry.hybridcloud.rpc_services.control_organization_provisioning import (
+    RpcOrganizationSlugReservation,
+)
 from sentry.services.hybrid_cloud.region import ByRegionName
 from sentry.services.hybrid_cloud.rpc import RpcService, regional_rpc_method
 from sentry.services.organization import OrganizationProvisioningOptions
 from sentry.silo import SiloMode
 
 
-class OrganizationProvisioningRegionService(RpcService):
+class RegionOrganizationProvisioningRpcService(RpcService):
     """
     RPC Service class containing methods for provisioning an organization that
     has already reserved a global slug in the control silo. This is only
@@ -46,17 +48,29 @@ class OrganizationProvisioningRegionService(RpcService):
         region_name: str,
         org_slug_temporary_alias_res: RpcOrganizationSlugReservation,
     ) -> bool:
+        """
+        CAUTION: THIS IS ONLY INTENDED TO BE USED BY THE `organization_provisioning` RPC SERVICE.
+        DO NOT USE FOR LOCAL CHANGES.
+
+        An RPC method for processing a slug change on the region, after it has been reserved
+        as a temporary alias in the control silo.
+
+        :param region_name: The region name where the organization resides.
+        :param org_slug_temporary_alias_res: OrganizationSlugReservation for the new temporary alias.
+        :return: True if provisioning succeeded, False if a conflict occurred
+        """
         pass
 
     @classmethod
     def get_local_implementation(cls) -> RpcService:
-        from sentry.hybridcloud.rpc_services.organization_provisioning_region.impl import (
-            DatabaseBackedOrganizationProvisioningRegionService,
+        from sentry.hybridcloud.rpc_services.region_organization_provisioning.impl import (
+            DatabaseBackedRegionOrganizationProvisioningRpcService,
         )
 
-        return DatabaseBackedOrganizationProvisioningRegionService()
+        return DatabaseBackedRegionOrganizationProvisioningRpcService()
 
 
-organization_provisioning_region_service: OrganizationProvisioningRegionService = cast(
-    OrganizationProvisioningRegionService, OrganizationProvisioningRegionService.create_delegation()
+region_organization_provisioning_rpc_service: RegionOrganizationProvisioningRpcService = cast(
+    RegionOrganizationProvisioningRpcService,
+    RegionOrganizationProvisioningRpcService.create_delegation(),
 )
